Refactored code by adding a deck class and moving deck methods to it, added static card creation method to card class
