+++ src/main/java/controllers/ApplicationController.java
@@ -33,10 +33,7 @@ public class ApplicationController {
     
     public Result gameGet(){
         Game g = new Game();
-        g.buildDeck();
-        g.shuffle();
         g.dealFour();
-
         return Results.json().render(g);
     }
